fix(config): use u32 for ProfileInfo.context_window to satisfy specta BigInt restriction
Specta forbids exporting u64 types to TypeScript to avoid precision loss.
Context window values fit comfortably in u32 (max ~4B tokens).

+++ crates/agent-config/src/types/profile.rs
@@ -90,7 +90,7 @@ pub struct ProfileInfo {
     #[serde(default)]
     pub model_display: String,
     #[serde(default)]
-    pub context_window: Option<u64>,
+    pub context_window: Option<u32>,
     #[serde(default)]
     pub supports_vision: bool,
     #[serde(default)]
